Full implementation of the new export command
This command aims to solve the issue regarding the archive server eventually becoming full.

It solve this by exporting the channel messages to a file and then using a webpage to view the channel messages in it.

The webpage still needs to be worked on but its base structure is already set on stone